♻️ Refactor: modularize document anonymization
