+++ aymurai/text/anonymization/__init__.py
@@ -1,0 +1,7 @@
+from aymurai.text.anonymization.alignment import replace_labels_in_text
+from aymurai.text.anonymization.doc_anonymizer import DocAnonymizer
+
+__all__ = [
+    "DocAnonymizer",
+    "replace_labels_in_text",
+]
